add basic test for wls.Point

+++ src/main/scala/wls/Wls.scala
@@ -1,7 +1,7 @@
 package wls
 
 class Wls(val xPoints: Array[Double], val yPoints: Array[Double], val weights: Array[Double]) {
-  def this(xPoints: Array[Double], yPoints: Array[Double], weight: Int) =
+  def this(xPoints: Array[Double], yPoints: Array[Double], weight: Double) =
     this(xPoints, yPoints, Array.fill(xPoints.length)(weight))
   def this(xPoints: Array[Double], yPoints: Array[Double]) =
     this(xPoints, yPoints, 1)
